build.gradle: add an option to use the local libvlcjni code
The option is on by default: we use the local libvlcjni out of the box if it's present.
Using the published libvlcjni may only be useful for releases.

This can be turned off with a gradle command line option: -PuseLocalLibvlcJNI=false or by not having a libvlcjni folder in the root.

For now release and vlcBundle builds always use the published AAR.